@@ -26,7 +26,7 @@ install_requires =
     requests
     snowflake-connector-python>=2.8,<4.0
     structlog~=24.1.0
-python_requires = >=3.9
+python_requires = >=3.10
 include_package_data = True
 
 [options.entry_points]
